LPS-66923 Move MB related data handlers to the correct package
