+++ modules/apps/collaboration/message-boards/message-boards-service/src/main/java/com/liferay/message/boards/internal/exportimport/data/handler/MBThreadFlagStagedModelDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.internal.lar;
+package com.liferay.message.boards.internal.exportimport.data.handler;
 
 import com.liferay.exportimport.kernel.lar.ExportImportPathUtil;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
