+++ modules/apps/collaboration/message-boards/message-boards-web/src/main/java/com/liferay/message/boards/web/internal/exportimport/data/handler/MBPortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.message.boards.web.lar;
+package com.liferay.message.boards.web.internal.exportimport.data.handler;
 
 import com.liferay.exportimport.kernel.lar.BasePortletDataHandler;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
